fix: Resolve build errors and all 127 tests now pass
Build fixes:
- Update SqlParser package 0.5.0 -> 1.0.2
- Fix 'when' keyword collision in pattern matching (C# contextual keyword)
- Resolve namespace collisions by moving Series/DataFrame to root namespace
- Add type aliases in Polaire.cs to avoid method/type shadowing
- Create VectorBinaryOp delegate for Span generic type constraints
- Fix type conversions (long->int casts, DateTime, Memory<T>)
- Resolve Array ambiguity (Apache.Arrow vs System.Array)
- Suppress CS1591 XML documentation warnings
- Add missing 'using Xunit' to test files

Bug fixes:
- Fix AnyValue.InvalidCast error message (was showing expected twice)
- Fix GroupBy aggregation type inference for aliased expressions

Also added:
- Comprehensive README with examples and documentation
- MIT LICENSE file

+++ src/Polaire/Compute/DateTimeOperations.cs
@@ -4,7 +4,7 @@
 using Apache.Arrow;
 using Polaire.DataTypes;
 using Polaire.Core;
-using Polaire.Series;
+
 
 namespace Polaire.Compute;
 
@@ -63,12 +63,12 @@ public sealed class DateTimeOperations
 
     public Series Microsecond()
     {
-        return ExtractComponent(dt => (dt.Ticks / 10) % 1000);
+        return ExtractComponent(dt => (int)((dt.Ticks / 10) % 1000));
     }
 
     public Series Nanosecond()
     {
-        return ExtractComponent(dt => (dt.Ticks % 10) * 100);
+        return ExtractComponent(dt => (int)((dt.Ticks % 10) * 100));
     }
 
     // ============================================================================
